Fixes #116 - some base rendering stuff
* create base renderer class. base class for gedrenderer. does not
  know about ged objects
* remove unused section renderers
* clean up some PMD warnings
* *did not yet use base renderer outside of ged object rendering*
