+++ gedbrowser-renderer/src/main/java/org/schoellerfamily/gedbrowser/renderer/SourceRenderer.java
@@ -26,7 +26,6 @@ public final class SourceRenderer extends GedRenderer<Source> {
             final RenderingContext renderingContext,
             final CalendarProvider provider) {
         super(gedObject, rendererFactory, renderingContext, provider);
-        setSectionRenderer(new SourceSectionRenderer(this));
     }
 
     /**
