+++ gedbrowser-renderer/src/main/java/org/schoellerfamily/gedbrowser/renderer/SubmittorRenderer.java
@@ -21,6 +21,5 @@ public final class SubmittorRenderer extends GedRenderer<Submittor> {
             final RenderingContext renderingContext,
             final CalendarProvider provider) {
         super(gedObject, rendererFactory, renderingContext, provider);
-        setSectionRenderer(new SubmittorSectionRenderer(this));
     }
 }
