+++ gedbrowser-renderer/src/test/java/org/schoellerfamily/gedbrowser/renderer/test/AttributeRendererTest.java
@@ -11,7 +11,6 @@ import org.schoellerfamily.gedbrowser.renderer.ApplicationInfo;
 import org.schoellerfamily.gedbrowser.renderer.AttributeListItemRenderer;
 import org.schoellerfamily.gedbrowser.renderer.AttributePhraseRenderer;
 import org.schoellerfamily.gedbrowser.renderer.AttributeRenderer;
-import org.schoellerfamily.gedbrowser.renderer.AttributeSectionRenderer;
 import org.schoellerfamily.gedbrowser.renderer.GedRendererFactory;
 import org.schoellerfamily.gedbrowser.renderer.NullNameHtmlRenderer;
 import org.schoellerfamily.gedbrowser.renderer.NullNameIndexRenderer;
@@ -104,20 +103,6 @@ public final class AttributeRendererTest {
         assertTrue("Wrong renderer type",
                 renderer.getPhraseRenderer()
                 instanceof AttributePhraseRenderer);
-    }
-
-    /**
-     * Test that we are using the appropriate sub-renderers.
-     * We will test the sub-renderers directly.
-     */
-    @Test
-    public void testSectionRenderer() {
-        final AttributeRenderer renderer = new AttributeRenderer(
-                new Attribute(null), new GedRendererFactory(),
-                anonymousContext, provider);
-        assertTrue("Wrong renderer type",
-                renderer.getSectionRenderer()
-                instanceof AttributeSectionRenderer);
     }
     // TODO test render as page and renderAsListItem
 }
