+++ gedbrowser-renderer/src/test/java/org/schoellerfamily/gedbrowser/renderer/test/GedRendererTest.java
@@ -21,7 +21,6 @@ import org.schoellerfamily.gedbrowser.renderer.NullListItemRenderer;
 import org.schoellerfamily.gedbrowser.renderer.NullNameHtmlRenderer;
 import org.schoellerfamily.gedbrowser.renderer.NullNameIndexRenderer;
 import org.schoellerfamily.gedbrowser.renderer.NullPhraseRenderer;
-import org.schoellerfamily.gedbrowser.renderer.NullSectionRenderer;
 import org.schoellerfamily.gedbrowser.renderer.RenderingContext;
 import org.schoellerfamily.gedbrowser.renderer.SimpleAttributeListOpenRenderer;
 
@@ -105,18 +104,6 @@ public final class GedRendererTest {
         assertTrue("Wrong renderer type",
                 renderer.getPhraseRenderer()
                 instanceof NullPhraseRenderer);
-    }
-
-    /**
-     * Test that we are using the appropriate sub-renderers.
-     * We will test the sub-renderers directly.
-     */
-    @Test
-    public void testSectionRenderer() {
-        final GedRenderer<GedObject> renderer = createRenderer();
-        assertTrue("Wrong renderer type",
-                renderer.getSectionRenderer()
-                instanceof NullSectionRenderer);
     }
 
     /**
