+++ gedbrowser-renderer/src/test/java/org/schoellerfamily/gedbrowser/renderer/test/NullRendererTest.java
@@ -20,7 +20,6 @@ import org.schoellerfamily.gedbrowser.renderer.NullNameHtmlRenderer;
 import org.schoellerfamily.gedbrowser.renderer.NullNameIndexRenderer;
 import org.schoellerfamily.gedbrowser.renderer.NullPhraseRenderer;
 import org.schoellerfamily.gedbrowser.renderer.NullRenderer;
-import org.schoellerfamily.gedbrowser.renderer.NullSectionRenderer;
 import org.schoellerfamily.gedbrowser.renderer.RenderingContext;
 import org.schoellerfamily.gedbrowser.renderer.SimpleAttributeListOpenRenderer;
 
@@ -103,17 +102,6 @@ public final class NullRendererTest {
         final NullRenderer renderer = createRenderer();
         assertTrue("Wrong renderer type",
                 renderer.getPhraseRenderer() instanceof NullPhraseRenderer);
-    }
-
-    /**
-     * Test that we are using the appropriate sub-renderers. We will test the
-     * sub-renderers directly.
-     */
-    @Test
-    public void testSectionRenderer() {
-        final NullRenderer renderer = createRenderer();
-        assertTrue("Wrong renderer type",
-                renderer.getSectionRenderer() instanceof NullSectionRenderer);
     }
 
     /** */
