+++ gedbrowser-renderer/src/test/java/org/schoellerfamily/gedbrowser/renderer/test/SourceRendererTest.java
@@ -22,7 +22,6 @@ import org.schoellerfamily.gedbrowser.renderer.NullPhraseRenderer;
 import org.schoellerfamily.gedbrowser.renderer.RenderingContext;
 import org.schoellerfamily.gedbrowser.renderer.SimpleAttributeListOpenRenderer;
 import org.schoellerfamily.gedbrowser.renderer.SourceRenderer;
-import org.schoellerfamily.gedbrowser.renderer.SourceSectionRenderer;
 
 /**
  * @author Dick Schoeller
@@ -100,18 +99,6 @@ public final class SourceRendererTest {
         assertTrue("Wrong renderer type",
                 renderer.getPhraseRenderer()
                 instanceof NullPhraseRenderer);
-    }
-
-    /**
-     * Test that we are using the appropriate sub-renderers.
-     * We will test the sub-renderers directly.
-     */
-    @Test
-    public void testSectionRenderer() {
-        final SourceRenderer renderer = createRenderer();
-        assertTrue("Wrong renderer type",
-                renderer.getSectionRenderer()
-                instanceof SourceSectionRenderer);
     }
 
     /**
